+++ src/main/java/com/csci334/TimsHobbyShop/repository/Item_Repository.java
@@ -3,5 +3,5 @@ package com.csci334.TimsHobbyShop.repository;
 import com.csci334.TimsHobbyShop.model.Item;
 import org.springframework.data.repository.CrudRepository;
 
-public interface Item_Repository extends CrudRepository<Item, Integer> {
+public interface Item_Repository extends CrudRepository<Item, Long> {
 }
