+++ src/main/java/com/csci334/TimsHobbyShop/repository/SaleLineItem_Repository.java
@@ -3,5 +3,5 @@ package com.csci334.TimsHobbyShop.repository;
 import com.csci334.TimsHobbyShop.model.SaleLineItem;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SaleLineItem_Repository extends CrudRepository<SaleLineItem, Integer> {
+public interface SaleLineItem_Repository extends CrudRepository<SaleLineItem, Long> {
 }
