+++ src/main/java/com/csci334/TimsHobbyShop/repository/Sale_Repository.java
@@ -3,5 +3,5 @@ package com.csci334.TimsHobbyShop.repository;
 import com.csci334.TimsHobbyShop.model.Sale;
 import org.springframework.data.repository.CrudRepository;
 
-public interface Sale_Repository extends CrudRepository<Sale, Integer> {
+public interface Sale_Repository extends CrudRepository<Sale, Long> {
 }
